+++ Arena/ArenaLobbyData.cs
@@ -103,6 +103,8 @@ namespace RainMeadow
 
             [OnlineField(group = "arenaSetup")]
             public bool voidMasterEnabled;
+            [OnlineFieldHalf(group = "arenaSetup")]
+            public float voidSpawnLethalityFactor;
 
             [OnlineField(group = "arenaSetup")]
             public int amoebaDuration;
@@ -155,6 +157,9 @@ namespace RainMeadow
 
             [OnlineField(group = "arenaSetup")]
             public bool challengeDenEjection;
+
+            [OnlineField(group = "arenaSetup")]
+            public bool enableMeadowCosmetics;
 
             [OnlineField(group = "arenaSetup")]
             public int artiExplosionCapacity;
@@ -272,6 +277,7 @@ namespace RainMeadow
 
                 bannedSlugs = new(arena.bannedSlugs);
                 voidMasterEnabled = arena.voidMasterEnabled;
+                voidSpawnLethalityFactor = arena.voidSpawnLethalityFactor;
                 amoebaDuration = arena.amoebaDuration;
                 amoebaControl = arena.amoebaControl;
                 friendlyFire = arena.friendlyFire;
@@ -291,6 +297,8 @@ namespace RainMeadow
                 artiExplosionCapacity = arena.artiExplosionCount;
                 artiParryDistance = arena.artiParryDistanceMult;
                 artiParryLeniency = arena.artiParryLeniency;
+                enableMeadowCosmetics = arena.enableMeadowCosmetics;
+
             }
 
             public override void ReadTo(OnlineResource.ResourceData data, OnlineResource resource)
@@ -366,6 +374,8 @@ playerNumberWithTrophiesPerRound;
                 (lobby.gameMode as ArenaOnlineGameMode).bannedSlugs = bannedSlugs;
                 (lobby.gameMode as ArenaOnlineGameMode).piggyBack = piggyBack;
                 (lobby.gameMode as ArenaOnlineGameMode).voidMasterEnabled = voidMasterEnabled;
+                (lobby.gameMode as ArenaOnlineGameMode).voidSpawnLethalityFactor = voidSpawnLethalityFactor;
+
                 (lobby.gameMode as ArenaOnlineGameMode).amoebaDuration = amoebaDuration;
                 (lobby.gameMode as ArenaOnlineGameMode).amoebaControl = amoebaControl;
                 (lobby.gameMode as ArenaOnlineGameMode).friendlyFire = friendlyFire;
@@ -387,6 +397,8 @@ playerNumberWithTrophiesPerRound;
                 (lobby.gameMode as ArenaOnlineGameMode).artiExplosionCount = artiExplosionCapacity;
                 (lobby.gameMode as ArenaOnlineGameMode).artiParryDistanceMult = artiParryDistance;
                 (lobby.gameMode as ArenaOnlineGameMode).artiParryLeniency = artiParryLeniency;
+                (lobby.gameMode as ArenaOnlineGameMode).enableMeadowCosmetics = enableMeadowCosmetics;
+
             }
 
             public override Type GetDataType() => typeof(ArenaLobbyData);